pyproject.toml
[tool.poetry]
name = "excel-migration-framework"
version = "0.1.0"
description = "A framework for migrating Excel data using configurable rules and LLM integration"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
openpyxl = "^3.1.2"
langchain = "^0.1.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
